fixes window.ipfs is added after page load
