+++ add-on/src/lib/ipfs-proxy/acl-whitelist.json
@@ -1,0 +1,30 @@
+[
+  "block.get",
+  "block.stat",
+  "dag.get",
+  "dag.tree",
+  "dht.get",
+  "dht.findprovs",
+  "dht.findpeer",
+  "dht.query",
+  "files.cat",
+  "files.catPullStream",
+  "files.get",
+  "files.getReadableStream",
+  "files.getPullStream",
+  "files.ls",
+  "files.lsReadableStream",
+  "files.lsPullStream",
+  "object.get",
+  "object.data",
+  "object.links",
+  "object.stat",
+  "pin.ls",
+  "pubsub.subscribe",
+  "pubsub.unsubscribe",
+  "pubsub.peers",
+  "pubsub.ls",
+  "swarm.peers",
+  "swarm.addrs",
+  "swarm.localAddrs"
+]
