+++ package.json
@@ -17,15 +17,17 @@
     "build:copy:wx-polyfill-lib": "shx cp node_modules/webextension-polyfill/dist/browser-polyfill.min.js add-on/dist/contentScripts/browser-polyfill.min.js",
     "build:js": "browserify -p prundupify -t browserify-css -t [ browserify-package-json --global ] add-on/src/background/background.js add-on/src/options/options.js add-on/src/popup/browser-action/index.js add-on/src/popup/quick-upload.js add-on/src/pages/proxy-acl/index.js -p [ factor-bundle -o add-on/dist/background/background.js -o add-on/dist/options/options.js -o add-on/dist/popup/browser-action/browser-action.js -o add-on/dist/popup/quick-upload.js -o add-on/dist/pages/proxy-acl/proxy-acl.js ] -o add-on/dist/ipfs-companion-common.js",
     "build:content-scripts": "run-p build:content-scripts:**",
-    "build:content-scripts:ipfs-proxy:content": "browserify -s IpfsProxyContent add-on/src/contentScripts/ipfs-proxy/content.js -o add-on/dist/contentScripts/ipfs-proxy/content.js",
-    "build:content-scripts:ipfs-proxy:page": "browserify -s IpfsProxyPage add-on/src/contentScripts/ipfs-proxy/page.js -o add-on/dist/contentScripts/ipfs-proxy/page.js",
+    "build:content-scripts:ipfs-proxy": "run-s build:content-scripts:ipfs-proxy:*",
+    "build:content-scripts:ipfs-proxy:page": "browserify add-on/src/contentScripts/ipfs-proxy/page.js -o add-on/dist/contentScripts/ipfs-proxy/page.js",
+    "build:content-scripts:ipfs-proxy:content": "browserify -t brfs -s IpfsProxyContent add-on/src/contentScripts/ipfs-proxy/content.js -o add-on/dist/contentScripts/ipfs-proxy/content.js",
     "build:minimize-dist": "shx rm -rf add-on/dist/lib",
     "build:bundle-extension": "web-ext build -s add-on/ -i src/ -a build/",
-    "watch": "run-p watch:*",
+    "watch": "npm-run-all build:copy --parallel watch:*",
     "watch:js": "watchify -p prundupify -t browserify-css -t [ browserify-package-json --global ] add-on/src/background/background.js add-on/src/options/options.js add-on/src/popup/browser-action/index.js add-on/src/popup/quick-upload.js add-on/src/pages/proxy-acl/index.js -p [ factor-bundle -o add-on/dist/background/background.js -o add-on/dist/options/options.js -o add-on/dist/popup/browser-action/browser-action.js -o add-on/dist/popup/quick-upload.js -o add-on/dist/pages/proxy-acl/proxy-acl.js ] -o add-on/dist/ipfs-companion-common.js -v",
     "watch:content-scripts": "run-p watch:content-scripts:**",
-    "watch:content-scripts:ipfs-proxy:content": "watchify -s IpfsProxyContent add-on/src/contentScripts/ipfs-proxy/content.js -o add-on/dist/contentScripts/ipfs-proxy/content.js",
-    "watch:content-scripts:ipfs-proxy:page": "watchify -s IpfsProxyPage add-on/src/contentScripts/ipfs-proxy/page.js -o add-on/dist/contentScripts/ipfs-proxy/page.js",
+    "watch:content-scripts:ipfs-proxy": "run-s watch:content-scripts:ipfs-proxy:*",
+    "watch:content-scripts:ipfs-proxy:page": "watchify add-on/src/contentScripts/ipfs-proxy/page.js -o add-on/dist/contentScripts/ipfs-proxy/page.js -v",
+    "watch:content-scripts:ipfs-proxy:content": "nodemon --exec \"browserify -t brfs -s IpfsProxyContent add-on/src/contentScripts/ipfs-proxy/content.js -o add-on/dist/contentScripts/ipfs-proxy/content.js -v\" --watch add-on/src/contentScripts/ipfs-proxy/content.js --watch add-on/dist/contentScripts/ipfs-proxy/page.js",
     "test": "run-s test:*",
     "test:functional": " nyc --reporter=lcov --reporter=text mocha --require ignore-styles 'test/functional/**/*.test.js'",
     "lint": "run-s lint:*",
@@ -43,6 +45,7 @@
     "babel-preset-es2015": "6.24.1",
     "babel-preset-es2017": "6.24.1",
     "babelify": "8.0.0",
+    "brfs": "^1.4.4",
     "browserify": "14.5.0",
     "browserify-css": "0.14.0",
     "browserify-package-json": "1.0.1",
@@ -56,6 +59,7 @@
     "ignore-styles": "5.0.1",
     "mem-storage-area": "1.0.3",
     "mocha": "4.0.1",
+    "nodemon": "^1.14.11",
     "npm-run-all": "4.1.2",
     "nyc": "^11.4.1",
     "shx": "0.2.2",
@@ -71,7 +75,7 @@
     "file-type": "7.3.0",
     "ipfs": "0.27.6",
     "ipfs-api": "17.2.7",
-    "ipfs-postmsg-proxy": "2.1.0",
+    "ipfs-postmsg-proxy": "2.2.0",
     "is-ipfs": "0.3.2",
     "is-svg": "2.1.0",
     "lru_map": "0.3.3",
